Feat: UI 컴포넌트 데이터 제공 utils 로직 분리
